apps/api-gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
pyjwt
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
